refactor: Transformation du composant Microsoft en bouton de navigation avec dropdown
- Remplacement de MicrosoftConnectCard par MicrosoftConnectButton
- Intégration du bouton Microsoft dans la barre de navigation
- Logo Microsoft adaptatif : gris si déconnecté, coloré si connecté
- Indicateur visuel avec badge vert et bordure verte quand connecté
- Dropdown menu avec état détaillé (profil utilisateur, email, actions)
- Interface plus compacte et accessible depuis toute l'application
- Suppression de l'ancien composant card volumineux du dashboard
- Support complet du thème sombre et animations cohérentes

Actions de refactorisation:
- ✅ Création components/microsoft-connect-button.tsx
- ✅ Suppression components/dashboard/microsoft-connect-card.tsx
- ✅ Intégration dans components/layout/navigation.tsx
- ✅ Nettoyage des imports dans app/dashboard/page.tsx
- ✅ Positionnement optimal entre theme toggle et avatar utilisateur
